requirements.txt
django
django-compressor==1.4